cleaned up all files, renamed files to appropriate names.
