+++ src/react-modules/credentials/CredentialsScreen.js
@@ -1,32 +1,23 @@
 import React, { useState, useEffect } from "react";
 import "./styles/credentials.css";
-import { JSONUploadButton } from "../JSONUpload";
-import { NoSelection, Form, SelectionOverview } from "./components/index.js";
-import { CredentialsList, Edit } from "./index.js";
-import { useImmer } from "use-immer";
-import { LocalTextInput } from "../LocalTextInput";
-import { Credential2 } from "./components/Credential2";
+import { CredentialsForm } from "./CredentialsForm";
+import { CredentialsEdit } from "./CredentialsEdit";
+import { CredentialsList } from "./components/CredentialsList";
+import { NoSelection } from "../credentials/NoSelection";
 
-function Credentials({ selectedCharacter, setSelectedIndex, selectedIndex }) {
+function CredentialsScreen({
+  selectedCharacter,
+  setSelectedIndex,
+  selectedIndex,
+}) {
   const [componentToMount, setComponentToMount] = useState("no selection");
   const [selectedCredential, setSelectedCredential] = useState(0);
 
   const [formData, setFormData] = useState([]);
   const [tempData, setTempData] = useState([]);
 
-  const [showForm, setShowForm] = useState(false);
-
   // Check if the create button has been clicked to ensure that you cant spam the button.
   const [createButtonClicked, setCreateButtonClicked] = useState(false);
-
-  // useEffect(() => {
-  //   console.log("Your tempData array is: ", tempData);
-  //   console.log("Your formData array is: ", formData);
-  // }, [tempData, formData]);
-
-  // useEffect(() => {
-  //   console.log(`your index is currently ${selectedCredential}`);
-  // });
 
   const showMeMyJSON = () => {
     console.log("Your current formData JSON is: ", formData);
@@ -93,7 +84,6 @@ function Credentials({ selectedCharacter, setSelectedIndex, selectedIndex }) {
 
   // Create a credential with an empty object.
   const handleCreateButtonClick = (e) => {
-    // if (!createButtonClicked) {
     setSelectedCredential(formData.length);
     setTempData([
       ...tempData,
@@ -105,7 +95,6 @@ function Credentials({ selectedCharacter, setSelectedIndex, selectedIndex }) {
       },
     ]);
     setComponentToMount(e.target.getAttribute("data-button-id").split("-")[0]);
-    // }
     setCreateButtonClicked(true);
   };
 
@@ -131,7 +120,7 @@ function Credentials({ selectedCharacter, setSelectedIndex, selectedIndex }) {
       //   );
       case "create":
         return (
-          <Form
+          <CredentialsForm
             handleChange={handleChange(selectedCredential)}
             tempData={tempData}
             setTempData={setTempData}
@@ -143,7 +132,7 @@ function Credentials({ selectedCharacter, setSelectedIndex, selectedIndex }) {
         );
       case "edit":
         return (
-          <Edit
+          <CredentialsEdit
             handleChange={handleChange(selectedCredential)}
             formData={formData}
             tempData={tempData}
@@ -168,9 +157,6 @@ function Credentials({ selectedCharacter, setSelectedIndex, selectedIndex }) {
         CLEAR JSON
       </button>
       <div className=" two-column-container mx-20 my-16">
-
-
-        
         <div className="two-column-col md:w-3/5 pr-4">
           <div className="flex justify-between">
             <div>
@@ -214,12 +200,9 @@ function Credentials({ selectedCharacter, setSelectedIndex, selectedIndex }) {
           {/* <Credential2 /> */}
         </div>
 
-
         <div className="two-column-col md:w-2/5 bg-gray-300 p-6 rounded-md right-col">
           {renderComponent(componentToMount)}
         </div>
-
-
 
         <div className="flex mt-5 w-full justify-end ">
           {(componentToMount === "edit" || componentToMount === "create") && (
@@ -239,4 +222,4 @@ function Credentials({ selectedCharacter, setSelectedIndex, selectedIndex }) {
   );
 }
 
-export { Credentials };
+export { CredentialsScreen };
