+++ src/react-modules/credentials/components/Credential.js
@@ -1,4 +1,6 @@
 import React from "react";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faTrash } from "@fortawesome/free-solid-svg-icons";
 
 function Credential({
   issuerName,
@@ -6,32 +8,52 @@ function Credential({
   attributeCount,
   index,
   handleClick,
+  formData,
+  setFormData,
+  setTempData,
+  selectedCredential,
+  setSelectedCredential,
 }) {
+  const handleCredentialRemoval = (index) => {
+    if (formData.length === 1) return;
+
+    setTempData((prevData) => {
+      const newData = [...prevData];
+      newData.splice(index, 1);
+      return newData;
+    });
+
+    setFormData((prevData) => {
+      const newData = [...prevData];
+      newData.splice(index, 1);
+      return newData;
+    });
+
+    setSelectedCredential((prevVal) => (prevVal - 1 === 0 ? prevVal - 1 : 0));
+  };
   return (
     <>
-      <div data-cred-id={index} onClick={handleClick}>
-        <div className="credential-card p-4">
-          <div className="credential-card-heading">
-            <div className="credential-card-issuer-name">
-              <p className="text-slate-100 text-xs truncate">{issuerName}</p>
+      <div>
+        <div className="credential-container grid grid-cols-7">
+          <div
+            className="credential rounded col-span-6 grid grid-cols-4"
+            data-cred-id={index}
+            onClick={handleClick}
+          >
+            <div className="cred-image"></div>
+            <div className="cred-info col-span-2">
+              <p className="cred-issuer-name">{issuerName}</p>
+              <p className="cred-credential-name">{credentialName}</p>
             </div>
-            <p className="text-slate-200 text-xl font-bold truncate">
-              {credentialName}
+            <div className=" cred-attributes">
+              <p className="">Attributes: {attributeCount}</p>
+            </div>
+          </div>
+          {/* end of one credential  */}
+          <div className="cred-actions">
+            <p onClick={() => handleCredentialRemoval(index)}>
+              <FontAwesomeIcon icon={faTrash} />
             </p>
-            <div className="credential-card-avatar my-4 flex items-center justify-center">
-              <div className="credential-card-avatar-background">
-                <img
-                  className="h-auto max-w-full"
-                  src="./images/lawyer-icon.svg"
-                  alt=""
-                />
-              </div>
-            </div>
-            <div className="credential-card-credential-attributes mt-6">
-              <p className="text-slate-100 text-sm ">
-                Attributes:{attributeCount}
-              </p>
-            </div>
           </div>
         </div>
       </div>
