+++ src/react-modules/credentials/components/CredentialsList.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { useState } from "react";
 import { Credential } from "./Credential";
-import { Credential2 } from "./Credential2";
 
 function CredentialsList({
   setComponentToMount,
@@ -23,15 +22,7 @@ function CredentialsList({
       <div className="">
         {formData.length > 0 ? (
           formData.map((credential, index) => (
-            // <Credential
-            //   key={index}
-            //   index={index}
-            //   handleClick={handleClick}
-            //   issuerName={credential.issuer_name}
-            //   credentialName={credential.cred_name}
-            //   attributeCount={credential.attributes.length}
-            // />
-            <Credential2
+            <Credential
               key={index}
               index={index}
               handleClick={handleClick}
