--- src/react-modules/credentials/components/index.js
@@ -1,3 +1,0 @@
-export { Form } from "../Form.js";
-export { NoSelection } from "../NoSelection.js";
-export { SelectionOverview } from "../CredentialInfo.js";
