--- src/react-modules/credentials/index.js
@@ -1,3 +1,0 @@
-export { Credential } from "./components/Credential.js";
-export { CredentialsList } from "./components/CredentialsList.js";
-export { Edit } from "./Edit.js";
